pyproject.toml
[tool.poetry]
name = "mrfox"
version = "0.1.0"
description = "Telegramm bot assistant"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.5.1"
python-dotenv = "^0.12.0"
sqlalchemy = "^1.3.16"
requests = "^2.23.0"
deluge-client = "^1.8.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
